separate bin and pipeline core parts
